gemini_adapter_partial.py
import google.generativeai as genai
from google.generativeai.types import content_types
from collections import namedtuple
import json
import logging

logger = logging.getLogger(__name__)

# Mock classes to mimic OpenAI response structure
Message = namedtuple("Message", ["content", "tool_calls"])
Choice = namedtuple("Choice", ["message"])
Response = namedtuple("Response", ["choices"])
ToolCall = namedtuple("ToolCall", ["id", "function", "type"])
Function = namedtuple("Function", ["name", "arguments"])

class GeminiAdapter:
    def __init__(self, api_key, model_name="gemini-1.5-flash"):
        genai.configure(api_key=api_key)
        self.model = genai.GenerativeModel(model_name)
        self.chat = self
        self.completions = self

    def _convert_tools(self, tools):
        """Convert OpenAI tool schema to Gemini function declarations."""
        if not tools:
            return None
        
        gemini_tools = []
        for t in tools:
            if t['type'] == 'function':
                f = t['function']
                # Gemini expects a slightly different structure but often accepts the dict
                # We will construct a tool object
                gemini_tools.append(genai.protos.Tool(function_declarations=[
                    genai.protos.FunctionDeclaration(
                        name=f['name'],
                        description=f['description'],
                        parameters=f['parameters']
                    )
                ]))
        return gemini_tools

    def create(self, model, messages, tools=None, tool_choice=None):
        """Simulate openai.chat.completions.create"""
        
        # 1. Convert Messages to Gemini History
        chat_history = []
        system_instruction = None
        last_user_message = ""
        
        for msg in messages:
            role = msg["role"]
            content = msg.get("content", "")
            
            if role == "system":
                system_instruction = content
            elif role == "user":
                # Gemini doesn't like consecutive user messages usually, but logic here assumes alternation or last one is prompt
                # We'll just build the history.
                chat_history.append({"role": "user", "parts": [content]})
                last_user_message = content
            elif role == "assistant":
                # Handle tool calls in history if previously made
                parts = []
                if msg.get("tool_calls"):
                    for tc in msg["tool_calls"]:
                        # This is complex to reconstruct perfectly for Gemini history 
                        # without tracking the function call object itself.
                        # For simplicity in this stateless pass:
                        # We might skip complex tool history reconstruction if possible 
                        # or just append content if it exists.
                        pass
                
                if content:
                    chat_history.append({"role": "model", "parts": [content]})
            
            elif role == "tool":
                # This is a result from a tool
                # Gemini expects "function_response" part
                # Reconstructing exact flow is hard.
                # If we are in the "tool result" phase, we usually have:
                # User (Prompt) -> Model (ToolCall) -> Tool (Result) -> Model (Final Answer)
                
                # In the 'run' method of agent, we append messages.
                # If we see tool messages, it means we are in the second turn.
                
                # Let's simplify: 
                # If we provide a `chat_session`, we can send message.
                pass

        # If we have tools, config them.
        # But `genai` chat session handles history better.
        
        # ACTUALLY, simpler approach for single-turn logic:
        # Just extract the last user message and the history before it.
        # Current Agent implementation sends ALL messages every time.
        
        # Let's try to map the LAST message as the prompt, and previous as history.
        if messages[-1]['role'] == 'tool':
             # We are responding to a tool call.
             # This is where the adapter gets tricky.
             # We need to construct a proper FunctionResponse.
             pass
        
        # REFACTOR STRATEGY: 
        # Since Agent.py is relatively simple, maybe it's easier to modifying Agent.py 
        # to branch logic rather than perfect adapter.
        pass

# ... rethinking ...
