requirements.txt
streamlit
openai
beautifulsoup4
chromadb
sentence-transformers
html2text
requests
google-generativeai
